Create easystroke.appdata.xml
Hello,

I've been a long time fan of easystroke, ever since I got all the buttons on my MX 1000 and Performance MX mice working just the way I wanted them.

You may have learned about the upcoming software centers in GNOME and KDE. These require an appdata file [1], [2] to be installed in /usr/share/appdata/ similar to the way .desktop files are placed in /usr/share/applications/, e.g. if you have a "$(datadir)/applications" definition in your makefiles, you need to add a "$(datadir)/appdata" as well.

There is a pending proposal in Fedora to not include in the software center applications that do not provide such appdata.xml files, so I took the liberty to prepare one for easystroke -along with some screenshots- when I noticed it was on that list (I've been working with several upstream projects in the last few days to that end). Please consider using this file and screenshots or making your own, just make sure it passes validation. Also if you decide to use this file, take a quick look to see if I've missed something or if you'd rather use a different license.

Thanks for your great work!

[1] http://people.freedesktop.org/~hughsient/appdata/
[2] http://www.freedesktop.org/software/appstream/docs/chap-Quickstart.html